src/requirements.txt
ipython
requests
pillow
